+++ internal/builtins/collector/linux/procfs/cpu.go
@@ -55,10 +55,10 @@ func NewCPUCollector(cfgBaseName, procFSPath string) (collector.Collector, error
 
 	c := CPU{}
 	c.id = CPU_NAME
-	c.pkgID = PFS_PREFIX + c.id
+	c.pkgID = PKG_NAME + "." + c.id
+	c.logger = log.With().Str("pkg", PKG_NAME).Str("id", c.id).Logger()
 	c.procFSPath = procFSPath
 	c.file = filepath.Join(c.procFSPath, procFile)
-	c.logger = log.With().Str("pkg", c.pkgID).Logger()
 	c.metricStatus = map[string]bool{}
 	c.metricDefaultActive = true
 	c.baseTags = tags.FromList(tags.GetBaseTags())