+++ internal/builtins/collector/windows/wmi/cache.go
@@ -75,8 +75,8 @@ type cacheOptions struct {
 func NewCacheCollector(cfgBaseName string) (collector.Collector, error) {
 	c := Cache{}
 	c.id = "cache"
-	c.pkgID = "builtins.windows.wmi." + c.id
-	c.logger = log.With().Str("pkg", c.pkgID).Logger()
+	c.pkgID = PKG_NAME + "." + c.id
+	c.logger = log.With().Str("pkg", PKG_NAME).Str("id", c.id).Logger()
 	c.metricDefaultActive = true
 	c.metricNameChar = defaultMetricChar
 	c.metricNameRegex = defaultMetricNameRegex