+++ internal/builtins/collector/windows/wmi/collector.go
@@ -13,6 +13,7 @@ import (
 	"github.com/circonus-labs/circonus-agent/internal/builtins/collector"
 	cgm "github.com/circonus-labs/circonus-gometrics/v3"
 	"github.com/pkg/errors"
+	"github.com/rs/zerolog"
 )
 
 // Define stubs to satisfy the collector.Collector interface.
@@ -57,6 +58,11 @@ func (c *wmicommon) Inventory() collector.InventoryStats {
 		LastRunDuration: c.lastRunDuration.String(),
 		LastError:       c.lastError,
 	}
+}
+
+// Logger returns collector's instance of logger
+func (c *wmicommon) Logger() zerolog.Logger {
+	return c.logger
 }
 
 // cleanName is used to clean the metric name