+++ internal/builtins/collector/windows/wmi/objects.go
@@ -52,8 +52,8 @@ type objectsOptions struct {
 func NewObjectsCollector(cfgBaseName string) (collector.Collector, error) {
 	c := Objects{}
 	c.id = "objects"
-	c.pkgID = "builtins.windows.wmi." + c.id
-	c.logger = log.With().Str("pkg", c.pkgID).Logger()
+	c.pkgID = PKG_NAME + "." + c.id
+	c.logger = log.With().Str("pkg", PKG_NAME).Str("id", c.id).Logger()
 	c.metricDefaultActive = true
 	c.metricNameChar = defaultMetricChar
 	c.metricNameRegex = defaultMetricNameRegex